requirements.txt
requests
beautifulsoup4
pandas
python-telegram-bot
PyQt5
sqlite3
configparser